corrected community page
